pyproject.toml
[project]
name = "codenames-ai"
version = "0.1.0"
description = "AI agent that plays Code Names via embedding clusters and LLM reranking."
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "python-dotenv>=1.0",
    "wordfreq>=3.1",
    "spacy>=3.7",
    "pandas>=2.2",
    "pyarrow>=16.0",
    "numpy>=1.26",
    "fasttext-wheel>=0.9",
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
    "openai>=1.50",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codenames_ai"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"

[tool.ruff]
line-length = 100
target-version = "py312"
